trendsql/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg[binary]>=3.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.0.0
pytrends>=4.9.0
PyYAML>=6.0.1
pandas>=2.0.0
gunicorn>=21.2.0
